requirements.txt
flask==3.0.2
flask-cors==4.0.0
pymongo==4.6.2
bcrypt==4.1.2
python-dotenv==1.0.1